Integrated COVID19 data (#173)
* Integrated COVID19 data

Integrated the covid19 data into the front end. Data appears in a media card within the feed. Could potentially add functionality to make it stay in a fixed position on the screen

* Integrated COVID19 data

Integrated the covid19 data into the front end. Data appears in a media card within the feed. Could potentially add functionality to make it stay in a fixed position on the screen

* Covid19 filter bar

Added Covid19 as an option in the filter bar